Fix GPU rendering and remove non-existent save_stats calls
Fixes two issues preventing main_gpu.py from running:
1. TypeError: points must be number pairs (polygon rendering)
2. AttributeError: 'GPUEcosystem' object has no attribute 'save_stats'

**Issue 1: Polygon TypeError**

Problem:
  pygame.draw.polygon() requires point coordinates as Python floats/ints,
  not numpy scalar types. RiverGPU was passing numpy.float64 values.

Solution (src/river_gpu.py):
  - Explicitly convert polygon points to Python float:
    float(path_x[i] + perp_x * effective_half_width)
  - Applied to both river banks and island polygons
  - pygame now accepts the point format

**Issue 2: Missing save_stats Method**

Problem:
  main_gpu.py called ecosystem.save_stats() in 3 places, but
  GPUEcosystem doesn't have this method. The old GPUVisualizer
  class tracked stats internally and saved with np.savez().

Solution (src/main_gpu.py):
  - Removed all ecosystem.save_stats() calls
  - Removed autosave_interval parameter
  - Removed stats tracking code (can be added back later if needed)
  - Simplified to just run simulation and render
  - GPU version now mirrors CPU version structure

**Changes:**

1. src/river_gpu.py:
   - Convert polygon coordinates to Python float (lines 294-300, 352-355, 379-382)
   - Ensures pygame compatibility

2. src/main_gpu.py:
   - Removed autosave_interval variable
   - Removed ecosystem.save_stats() calls (3 locations)
   - Removed "S - Save statistics" from controls
   - Removed "Auto-save" message
   - Ignore save_requested from renderer.render()

**Testing:**
✅ GPU simulation runs without errors
✅ Polygon rendering works correctly
✅ No AttributeError on save_stats
✅ Renderer displays properly

Both CPU and GPU versions now work identically!

+++ src/river_gpu.py
@@ -290,14 +290,14 @@ class RiverGPU:
             else:
                 effective_half_width = self.river_width / 2
 
-            # Calculate bank positions
+            # Calculate bank positions (convert to Python float for pygame)
             river_left_bank.append((
-                path_x[i] + perp_x * effective_half_width,
-                path_y[i] + perp_y * effective_half_width
+                float(path_x[i] + perp_x * effective_half_width),
+                float(path_y[i] + perp_y * effective_half_width)
             ))
             river_right_bank.append((
-                path_x[i] - perp_x * effective_half_width,
-                path_y[i] - perp_y * effective_half_width
+                float(path_x[i] - perp_x * effective_half_width),
+                float(path_y[i] - perp_y * effective_half_width)
             ))
 
         # Create closed polygon: left bank forward, right bank backward
@@ -349,9 +349,10 @@ class RiverGPU:
 
             half_island = (self.island_width / 2) * taper
 
+            # Convert to Python float for pygame
             island_points.append((
-                path_x[i] + perp_x * half_island,
-                path_y[i] + perp_y * half_island
+                float(path_x[i] + perp_x * half_island),
+                float(path_y[i] + perp_y * half_island)
             ))
 
         # Generate points along bottom edge (right to left)
@@ -375,9 +376,10 @@ class RiverGPU:
             taper = np.sin(t_local * np.pi)
             half_island = (self.island_width / 2) * taper
 
+            # Convert to Python float for pygame
             island_points.append((
-                path_x[i] - perp_x * half_island,
-                path_y[i] - perp_y * half_island
+                float(path_x[i] - perp_x * half_island),
+                float(path_y[i] - perp_y * half_island)
             ))
 
         return island_points
